chore: use modal manager for command palette
